# 리팩토링

See merge request s10-webmobile2-sub2/S10P12A109!41

Former-commit-id: 273636b8f70d7fbacf90f8fa816b37f9e298d676

+++ BACK/muscle_maker/src/main/java/com/ssafy/muscle_maker/dto/clubs/request/ClubLeaderRequest.java
@@ -7,6 +7,6 @@ import lombok.*;
 @AllArgsConstructor
 @Builder
 public class ClubLeaderRequest {
-    private long leaderId;
-    private long memberId;
+    private int leaderId;
+    private int memberId;
 }
