+++ BACK/muscle_maker/src/main/java/com/ssafy/muscle_maker/dto/inBody/request/CreateInBodyRequest.java
@@ -8,8 +8,8 @@ import lombok.*;
 @NoArgsConstructor
 public class CreateInBodyRequest {
 
-    private Float weight;
-    private Float muscleMass;
-    private Float fatMass;
+    private Double weight;
+    private Double muscleMass;
+    private Double fatMass;
 
 }
