+++ BACK/muscle_maker/src/main/java/com/ssafy/muscle_maker/repository/BasketRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BasketRepository extends JpaRepository<Basket, Long> {
+public interface BasketRepository extends JpaRepository<Basket, Integer> {
     Basket findByUserAndFlagFalse(User user);
 }
